Move the types into their own modules
